+++ SwiftBook/ViewControllers/WallstreetbetListViewController.swift
@@ -35,22 +35,6 @@ final class WallstreetbetListViewController: UITableViewController {
         return cell
     }
     
-    // MARK: - Private Methods
-    
-    private func showAlert(status: StatusAlert) {
-        DispatchQueue.main.async {
-            let alert = UIAlertController(
-                title: status.title,
-                message: status.message,
-                preferredStyle: .alert
-            )
-            
-            let okAction = UIAlertAction(title: "OK", style: .default)
-            alert.addAction(okAction)
-            self.present(alert, animated: true)
-        }
-    }
-    
 }
 
 extension WallstreetbetListViewController {
@@ -66,7 +50,9 @@ extension WallstreetbetListViewController {
                         self?.tableView.reloadData()
                     }
                 case .failure(let error):
-                    self?.showAlert(status: .failed)
+                    if let self = self {
+                        ShowAlert.shared.showAlert(where: self, status: .failed)
+                    }
                     print(error)
                 }
             }
